+++ data-transfer/src/main/java/it/eng/datatransfer/rest/api/DataTransferAPIController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
+import it.eng.datatransfer.model.DataTransferRequest;
 import it.eng.datatransfer.service.DataTransferAPIService;
 import it.eng.tools.controller.ApiEndpoints;
 import it.eng.tools.model.DSpaceConstants;
@@ -38,13 +39,9 @@ public class DataTransferAPIController {
 	
 	
 	@PostMapping
-	public ResponseEntity<GenericApiResponse<JsonNode>> requestTransfer(@RequestBody JsonNode requestTransferRequest) {
-		String targetConnector = requestTransferRequest.get("Forward-To").asText();
-		String agreementId = requestTransferRequest.get("agreementId").asText();
-		String format = requestTransferRequest.get(DSpaceConstants.FORMAT).asText();
-    	JsonNode dataAddress = requestTransferRequest.get(DSpaceConstants.DATA_ADDRESS);
+	public ResponseEntity<GenericApiResponse<JsonNode>> requestTransfer(@RequestBody DataTransferRequest dataTransferRequest ) {
 		log.info("Consumer sends transfer request");
-		JsonNode response = apiService.requestTransfer(targetConnector, agreementId, format, dataAddress);
+		JsonNode response = apiService.requestTransfer(dataTransferRequest);
 		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON)
 				.body(GenericApiResponse.success(response, "Data transfer requested"));
 	}
@@ -59,9 +56,11 @@ public class DataTransferAPIController {
 	 */
 	@GetMapping(path = { "", "/{transferProcessId}" })
 	public ResponseEntity<GenericApiResponse<Collection<JsonNode>>> getTransfersProcess(
-			@PathVariable(required = false) String transferProcessId, @RequestParam(required = false) String state) {
-		log.info("Ferching transfer process id - {}, state {}", transferProcessId, state);
-		Collection<JsonNode> response = apiService.findDataTransfers(transferProcessId, state);
+			@PathVariable(required = false) String transferProcessId,
+			@RequestParam(required = false) String state,
+			@RequestParam(required = false) String role) {
+		log.info("Fetching transfer process id - {}, state {}", transferProcessId, state);
+		Collection<JsonNode> response = apiService.findDataTransfers(transferProcessId, state, role);
 		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON)
 				.body(GenericApiResponse.success(response, "Fetching transfer process"));
 	}