+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.1] - 21-11-2024
+## [0.1.1] - 22-11-2024
 
 ### Changed
 
@@ -11,6 +11,17 @@ All notable changes to this project will be documented in this file.
 ### Added
 
  - Provider handleTermination request
+
+## [0.1.1] - 21-11-2024
+
+### Added
+
+ - Added role filter to GET Transport Process API endpoint
+
+### Changed
+
+ - Initiate transfer process API endpoint now uses initialized transfer processes
+ - Download now uses data from DB (not the hardcoded John Doe)
 
 ## [0.1.1] - 19-11-2024